# backEnd/codingAddict/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "mern",
    "contact": {},
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "variables": {}
    }
  ],
  "paths": {
...
    "/api/v1/products": {
      "post": {
        "tags": [
          "Admin - Product"
        ],
        "summary": "createProducts",
        "operationId": "createProducts",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/createProductsRequest"
              },
              "example": {
                "name": "string",
                "price": "number",
                "description": "string",
                "image": "string",
                "category": "string",
                "company": "string",
                "colors": "string",
                "inventory": "number"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "tags": [
          "User - Product"
        ],
        "summary": "getAllProducts",
        "operationId": "getAllProducts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/products/{id}": {
      "patch": {
        "tags": [
          "Admin - Product"
        ],
        "summary": "updateProduct",
        "operationId": "updateProduct",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/updateProductRequest"
              },
              "example": {
                "name": "string",
                "price": "number",
                "description": "string",
                "image": "string",
                "category": "string",
                "company": "string",
                "colors": "string",
                "inventory": "number"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Admin - Product"
        ],
        "summary": "deleteProduct",
        "operationId": "deleteProduct",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "tags": [
          "User - Product"
        ],
        "summary": "getSingleProduct",
        "operationId": "getSingleProduct",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "createProductsRequest": {
        "title": "createProductsRequest",
        "required": [
          "name",
          "price",
          "description",
          "image",
          "category",
          "company",
          "colors",
          "inventory"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "price": {
            "type": "integer",
            "format": "int32"
          },
          "description": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "company": {
            "type": "string"
          },
          "colors": {
            "type": "string"
          },
          "inventory": {
            "type": "integer",
            "format": "int32"
          }
        },
        "example": {
          "name": "chair",
          "price": 3000,
          "description": "something something",
          "image": "/uploads/example.jpeg",
          "category": "office",
          "company": "liddy",
          "colors": "",
          "inventory": 15
        }
      },
      "updateProductRequest": {
        "title": "updateProductRequest",
        "required": [
          "name",
          "price",
          "description",
          "image",
          "category",
          "company",
          "colors",
          "inventory"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "price": {
            "type": "integer",
            "format": "int32"
          },
          "description": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "company": {
            "type": "string"
          },
          "colors": {
            "type": "string"
          },
          "inventory": {
            "type": "integer",
            "format": "int32"
          }
        },
        "example": {
          "name": "chair",
          "price": 3500,
          "description": "something something",
          "image": "/uploads/example.jpeg",
          "category": "office",
          "company": "liddy",
          "colors": "",
          "inventory": 15
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [],
...
}